chore: rename react components dragdrop to selction
